Simplify table-merge
The `filter` has been removed. Its main function was to transform data before and after it was passed to the `action`.
For example, when entering edit mode in an action, it would convert a serialized YAML string back to struct, and then convert it back after editing was complete.

It was removed because:
- The data structure seen in editer differs from that in batch mode or the database, causing confusion.
    - In most cases, the benefit was the ability to use syntax highlighting in the editor.
- It added unnecessary complexity.
- It impacted performance.